requirements.txt
py-cord
python-dotenv
SQLAlchemy
aiosqlite
PyNaCl
PyYAML
